+++ lib/src/classes.g.dart
@@ -7,9 +7,10 @@ part of 'classes.dart';
 // **************************************************************************
 
 UserInfo _$UserInfoFromJson(Map<String, dynamic> json) {
-  return UserInfo()
-    ..uid = json['uid'] as int
-    ..userAccount = json['userAccount'] as String;
+  return UserInfo(
+    json['uid'] as int,
+    json['userAccount'] as String,
+  );
 }
 
 Map<String, dynamic> _$UserInfoToJson(UserInfo instance) => <String, dynamic>{
@@ -19,16 +20,24 @@ Map<String, dynamic> _$UserInfoToJson(UserInfo instance) => <String, dynamic>{
 
 VideoDimensions _$VideoDimensionsFromJson(Map<String, dynamic> json) {
   return VideoDimensions(
-    json['width'] as int,
-    json['height'] as int,
+    width: json['width'] as int?,
+    height: json['height'] as int?,
   );
 }
 
-Map<String, dynamic> _$VideoDimensionsToJson(VideoDimensions instance) =>
-    <String, dynamic>{
-      'width': instance.width,
-      'height': instance.height,
-    };
+Map<String, dynamic> _$VideoDimensionsToJson(VideoDimensions instance) {
+  final val = <String, dynamic>{};
+
+  void writeNotNull(String key, dynamic value) {
+    if (value != null) {
+      val[key] = value;
+    }
+  }
+
+  writeNotNull('width', instance.width);
+  writeNotNull('height', instance.height);
+  return val;
+}
 
 VideoEncoderConfiguration _$VideoEncoderConfigurationFromJson(
     Map<String, dynamic> json) {
@@ -39,8 +48,8 @@ VideoEncoderConfiguration _$VideoEncoderConfigurationFromJson(
     frameRate: _$enumDecodeNullable(_$VideoFrameRateEnumMap, json['frameRate']),
     minFrameRate:
         _$enumDecodeNullable(_$VideoFrameRateEnumMap, json['minFrameRate']),
-    bitrate: json['bitrate'] as int,
-    minBitrate: json['minBitrate'] as int,
+    bitrate: json['bitrate'] as int?,
+    minBitrate: json['minBitrate'] as int?,
     orientationMode: _$enumDecodeNullable(
         _$VideoOutputOrientationModeEnumMap, json['orientationMode']),
     degradationPrefer: _$enumDecodeNullable(
@@ -73,36 +82,41 @@ Map<String, dynamic> _$VideoEncoderConfigurationToJson(
   return val;
 }
 
-T _$enumDecode<T>(
-  Map<T, dynamic> enumValues,
-  dynamic source, {
-  T unknownValue,
+K _$enumDecode<K, V>(
+  Map<K, V> enumValues,
+  Object? source, {
+  K? unknownValue,
 }) {
   if (source == null) {
-    throw ArgumentError('A value must be provided. Supported values: '
-        '${enumValues.values.join(', ')}');
+    throw ArgumentError(
+      'A value must be provided. Supported values: '
+      '${enumValues.values.join(', ')}',
+    );
   }
 
-  final value = enumValues.entries
-      .singleWhere((e) => e.value == source, orElse: () => null)
-      ?.key;
-
-  if (value == null && unknownValue == null) {
-    throw ArgumentError('`$source` is not one of the supported values: '
-        '${enumValues.values.join(', ')}');
-  }
-  return value ?? unknownValue;
+  return enumValues.entries.singleWhere(
+    (e) => e.value == source,
+    orElse: () {
+      if (unknownValue == null) {
+        throw ArgumentError(
+          '`$source` is not one of the supported values: '
+          '${enumValues.values.join(', ')}',
+        );
+      }
+      return MapEntry(unknownValue, enumValues.values.first);
+    },
+  ).key;
 }
 
-T _$enumDecodeNullable<T>(
-  Map<T, dynamic> enumValues,
+K? _$enumDecodeNullable<K, V>(
+  Map<K, V> enumValues,
   dynamic source, {
-  T unknownValue,
+  K? unknownValue,
 }) {
   if (source == null) {
     return null;
   }
-  return _$enumDecode<T>(enumValues, source, unknownValue: unknownValue);
+  return _$enumDecode<K, V>(enumValues, source, unknownValue: unknownValue);
 }
 
 const _$VideoFrameRateEnumMap = {
@@ -138,9 +152,9 @@ BeautyOptions _$BeautyOptionsFromJson(Map<String, dynamic> json) {
   return BeautyOptions(
     lighteningContrastLevel: _$enumDecodeNullable(
         _$LighteningContrastLevelEnumMap, json['lighteningContrastLevel']),
-    lighteningLevel: (json['lighteningLevel'] as num)?.toDouble(),
-    smoothnessLevel: (json['smoothnessLevel'] as num)?.toDouble(),
-    rednessLevel: (json['rednessLevel'] as num)?.toDouble(),
+    lighteningLevel: (json['lighteningLevel'] as num?)?.toDouble(),
+    smoothnessLevel: (json['smoothnessLevel'] as num?)?.toDouble(),
+    rednessLevel: (json['rednessLevel'] as num?)?.toDouble(),
   );
 }
 
@@ -170,41 +184,16 @@ const _$LighteningContrastLevelEnumMap = {
 AgoraImage _$AgoraImageFromJson(Map<String, dynamic> json) {
   return AgoraImage(
     json['url'] as String,
-    json['x'] as int,
-    json['y'] as int,
-    json['width'] as int,
-    json['height'] as int,
+    x: json['x'] as int?,
+    y: json['y'] as int?,
+    width: json['width'] as int?,
+    height: json['height'] as int?,
   );
 }
 
-Map<String, dynamic> _$AgoraImageToJson(AgoraImage instance) =>
-    <String, dynamic>{
-      'url': instance.url,
-      'x': instance.x,
-      'y': instance.y,
-      'width': instance.width,
-      'height': instance.height,
-    };
-
-TranscodingUser _$TranscodingUserFromJson(Map<String, dynamic> json) {
-  return TranscodingUser(
-    json['uid'] as int,
-    json['x'] as int,
-    json['y'] as int,
-    width: json['width'] as int,
-    height: json['height'] as int,
-    zOrder: json['zOrder'] as int,
-    alpha: (json['alpha'] as num)?.toDouble(),
-    audioChannel:
-        _$enumDecodeNullable(_$AudioChannelEnumMap, json['audioChannel']),
-  );
-}
-
-Map<String, dynamic> _$TranscodingUserToJson(TranscodingUser instance) {
+Map<String, dynamic> _$AgoraImageToJson(AgoraImage instance) {
   final val = <String, dynamic>{
-    'uid': instance.uid,
-    'x': instance.x,
-    'y': instance.y,
+    'url': instance.url,
   };
 
   void writeNotNull(String key, dynamic value) {
@@ -213,6 +202,40 @@ Map<String, dynamic> _$TranscodingUserToJson(TranscodingUser instance) {
     }
   }
 
+  writeNotNull('x', instance.x);
+  writeNotNull('y', instance.y);
+  val['width'] = instance.width;
+  val['height'] = instance.height;
+  return val;
+}
+
+TranscodingUser _$TranscodingUserFromJson(Map<String, dynamic> json) {
+  return TranscodingUser(
+    json['uid'] as int,
+    x: json['x'] as int?,
+    y: json['y'] as int?,
+    width: json['width'] as int?,
+    height: json['height'] as int?,
+    zOrder: json['zOrder'] as int?,
+    alpha: (json['alpha'] as num?)?.toDouble(),
+    audioChannel:
+        _$enumDecodeNullable(_$AudioChannelEnumMap, json['audioChannel']),
+  );
+}
+
+Map<String, dynamic> _$TranscodingUserToJson(TranscodingUser instance) {
+  final val = <String, dynamic>{
+    'uid': instance.uid,
+  };
+
+  void writeNotNull(String key, dynamic value) {
+    if (value != null) {
+      val[key] = value;
+    }
+  }
+
+  writeNotNull('x', instance.x);
+  writeNotNull('y', instance.y);
   writeNotNull('width', instance.width);
   writeNotNull('height', instance.height);
   writeNotNull('zOrder', instance.zOrder);
@@ -232,18 +255,16 @@ const _$AudioChannelEnumMap = {
 
 LiveTranscoding _$LiveTranscodingFromJson(Map<String, dynamic> json) {
   return LiveTranscoding(
-    (json['transcodingUsers'] as List)
-        ?.map((e) => e == null
-            ? null
-            : TranscodingUser.fromJson(e as Map<String, dynamic>))
-        ?.toList(),
-    width: json['width'] as int,
-    height: json['height'] as int,
-    videoBitrate: json['videoBitrate'] as int,
+    (json['transcodingUsers'] as List<dynamic>)
+        .map((e) => TranscodingUser.fromJson(e as Map<String, dynamic>))
+        .toList(),
+    width: json['width'] as int?,
+    height: json['height'] as int?,
+    videoBitrate: json['videoBitrate'] as int?,
     videoFramerate:
         _$enumDecodeNullable(_$VideoFrameRateEnumMap, json['videoFramerate']),
-    lowLatency: json['lowLatency'] as bool,
-    videoGop: json['videoGop'] as int,
+    lowLatency: json['lowLatency'] as bool?,
+    videoGop: json['videoGop'] as int?,
     watermark: json['watermark'] == null
         ? null
         : AgoraImage.fromJson(json['watermark'] as Map<String, dynamic>),
@@ -252,7 +273,7 @@ LiveTranscoding _$LiveTranscodingFromJson(Map<String, dynamic> json) {
         : AgoraImage.fromJson(json['backgroundImage'] as Map<String, dynamic>),
     audioSampleRate: _$enumDecodeNullable(
         _$AudioSampleRateTypeEnumMap, json['audioSampleRate']),
-    audioBitrate: json['audioBitrate'] as int,
+    audioBitrate: json['audioBitrate'] as int?,
     audioChannels:
         _$enumDecodeNullable(_$AudioChannelEnumMap, json['audioChannels']),
     audioCodecProfile: _$enumDecodeNullable(
@@ -261,7 +282,7 @@ LiveTranscoding _$LiveTranscodingFromJson(Map<String, dynamic> json) {
         _$VideoCodecProfileTypeEnumMap, json['videoCodecProfile']),
     backgroundColor: LiveTranscoding._$ColorFromJson(
         json['backgroundColor'] as Map<String, dynamic>),
-    userConfigExtraInfo: json['userConfigExtraInfo'] as String,
+    userConfigExtraInfo: json['userConfigExtraInfo'] as String?,
   );
 }
 
@@ -295,7 +316,7 @@ Map<String, dynamic> _$LiveTranscodingToJson(LiveTranscoding instance) {
       LiveTranscoding._$ColorToJson(instance.backgroundColor));
   writeNotNull('userConfigExtraInfo', instance.userConfigExtraInfo);
   val['transcodingUsers'] =
-      instance.transcodingUsers?.map((e) => e?.toJson())?.toList();
+      instance.transcodingUsers.map((e) => e.toJson()).toList();
   return val;
 }
 
@@ -318,14 +339,16 @@ const _$VideoCodecProfileTypeEnumMap = {
 
 ChannelMediaInfo _$ChannelMediaInfoFromJson(Map<String, dynamic> json) {
   return ChannelMediaInfo(
+    json['channelName'] as String,
     json['uid'] as int,
-    channelName: json['channelName'] as String,
-    token: json['token'] as String,
+    token: json['token'] as String?,
   );
 }
 
 Map<String, dynamic> _$ChannelMediaInfoToJson(ChannelMediaInfo instance) {
-  final val = <String, dynamic>{};
+  final val = <String, dynamic>{
+    'channelName': instance.channelName,
+  };
 
   void writeNotNull(String key, dynamic value) {
     if (value != null) {
@@ -333,7 +356,6 @@ Map<String, dynamic> _$ChannelMediaInfoToJson(ChannelMediaInfo instance) {
     }
   }
 
-  writeNotNull('channelName', instance.channelName);
   writeNotNull('token', instance.token);
   val['uid'] = instance.uid;
   return val;
@@ -342,22 +364,18 @@ Map<String, dynamic> _$ChannelMediaInfoToJson(ChannelMediaInfo instance) {
 ChannelMediaRelayConfiguration _$ChannelMediaRelayConfigurationFromJson(
     Map<String, dynamic> json) {
   return ChannelMediaRelayConfiguration(
-    json['srcInfo'] == null
-        ? null
-        : ChannelMediaInfo.fromJson(json['srcInfo'] as Map<String, dynamic>),
-    (json['destInfos'] as List)
-        ?.map((e) => e == null
-            ? null
-            : ChannelMediaInfo.fromJson(e as Map<String, dynamic>))
-        ?.toList(),
+    ChannelMediaInfo.fromJson(json['srcInfo'] as Map<String, dynamic>),
+    (json['destInfos'] as List<dynamic>)
+        .map((e) => ChannelMediaInfo.fromJson(e as Map<String, dynamic>))
+        .toList(),
   );
 }
 
 Map<String, dynamic> _$ChannelMediaRelayConfigurationToJson(
         ChannelMediaRelayConfiguration instance) =>
     <String, dynamic>{
-      'srcInfo': instance.srcInfo?.toJson(),
-      'destInfos': instance.destInfos?.map((e) => e?.toJson())?.toList(),
+      'srcInfo': instance.srcInfo.toJson(),
+      'destInfos': instance.destInfos.map((e) => e.toJson()).toList(),
     };
 
 LastmileProbeConfig _$LastmileProbeConfigFromJson(Map<String, dynamic> json) {
@@ -380,31 +398,40 @@ Map<String, dynamic> _$LastmileProbeConfigToJson(
 
 Rectangle _$RectangleFromJson(Map<String, dynamic> json) {
   return Rectangle(
-    json['x'] as int,
-    json['y'] as int,
-    json['width'] as int,
-    json['height'] as int,
+    x: json['x'] as int?,
+    y: json['y'] as int?,
+    width: json['width'] as int?,
+    height: json['height'] as int?,
   );
 }
 
-Map<String, dynamic> _$RectangleToJson(Rectangle instance) => <String, dynamic>{
-      'x': instance.x,
-      'y': instance.y,
-      'width': instance.width,
-      'height': instance.height,
-    };
+Map<String, dynamic> _$RectangleToJson(Rectangle instance) {
+  final val = <String, dynamic>{};
+
+  void writeNotNull(String key, dynamic value) {
+    if (value != null) {
+      val[key] = value;
+    }
+  }
+
+  writeNotNull('x', instance.x);
+  writeNotNull('y', instance.y);
+  writeNotNull('width', instance.width);
+  writeNotNull('height', instance.height);
+  return val;
+}
 
 WatermarkOptions _$WatermarkOptionsFromJson(Map<String, dynamic> json) {
   return WatermarkOptions(
-    json['positionInLandscapeMode'] == null
+    visibleInPreview: json['visibleInPreview'] as bool?,
+    positionInLandscapeMode: json['positionInLandscapeMode'] == null
         ? null
         : Rectangle.fromJson(
             json['positionInLandscapeMode'] as Map<String, dynamic>),
-    json['positionInPortraitMode'] == null
+    positionInPortraitMode: json['positionInPortraitMode'] == null
         ? null
         : Rectangle.fromJson(
             json['positionInPortraitMode'] as Map<String, dynamic>),
-    visibleInPreview: json['visibleInPreview'] as bool,
   );
 }
 
@@ -418,23 +445,25 @@ Map<String, dynamic> _$WatermarkOptionsToJson(WatermarkOptions instance) {
   }
 
   writeNotNull('visibleInPreview', instance.visibleInPreview);
-  val['positionInLandscapeMode'] = instance.positionInLandscapeMode?.toJson();
-  val['positionInPortraitMode'] = instance.positionInPortraitMode?.toJson();
+  writeNotNull(
+      'positionInLandscapeMode', instance.positionInLandscapeMode?.toJson());
+  writeNotNull(
+      'positionInPortraitMode', instance.positionInPortraitMode?.toJson());
   return val;
 }
 
 LiveInjectStreamConfig _$LiveInjectStreamConfigFromJson(
     Map<String, dynamic> json) {
   return LiveInjectStreamConfig(
-    width: json['width'] as int,
-    height: json['height'] as int,
-    videoGop: json['videoGop'] as int,
+    width: json['width'] as int?,
+    height: json['height'] as int?,
+    videoGop: json['videoGop'] as int?,
     videoFramerate:
         _$enumDecodeNullable(_$VideoFrameRateEnumMap, json['videoFramerate']),
-    videoBitrate: json['videoBitrate'] as int,
+    videoBitrate: json['videoBitrate'] as int?,
     audioSampleRate: _$enumDecodeNullable(
         _$AudioSampleRateTypeEnumMap, json['audioSampleRate']),
-    audioBitrate: json['audioBitrate'] as int,
+    audioBitrate: json['audioBitrate'] as int?,
     audioChannels:
         _$enumDecodeNullable(_$AudioChannelEnumMap, json['audioChannels']),
   );
@@ -466,19 +495,18 @@ Map<String, dynamic> _$LiveInjectStreamConfigToJson(
 CameraCapturerConfiguration _$CameraCapturerConfigurationFromJson(
     Map<String, dynamic> json) {
   return CameraCapturerConfiguration(
-    _$enumDecodeNullable(
+    preference: _$enumDecodeNullable(
         _$CameraCaptureOutputPreferenceEnumMap, json['preference']),
-    _$enumDecodeNullable(_$CameraDirectionEnumMap, json['cameraDirection']),
-    captureWidth: json['captureWidth'] as int,
-    captureHeight: json['captureHeight'] as int,
+    captureWidth: json['captureWidth'] as int?,
+    captureHeight: json['captureHeight'] as int?,
+    cameraDirection:
+        _$enumDecodeNullable(_$CameraDirectionEnumMap, json['cameraDirection']),
   );
 }
 
 Map<String, dynamic> _$CameraCapturerConfigurationToJson(
     CameraCapturerConfiguration instance) {
-  final val = <String, dynamic>{
-    'preference': _$CameraCaptureOutputPreferenceEnumMap[instance.preference],
-  };
+  final val = <String, dynamic>{};
 
   void writeNotNull(String key, dynamic value) {
     if (value != null) {
@@ -486,9 +514,12 @@ Map<String, dynamic> _$CameraCapturerConfigurationToJson(
     }
   }
 
+  writeNotNull('preference',
+      _$CameraCaptureOutputPreferenceEnumMap[instance.preference]);
   writeNotNull('captureWidth', instance.captureWidth);
   writeNotNull('captureHeight', instance.captureHeight);
-  val['cameraDirection'] = _$CameraDirectionEnumMap[instance.cameraDirection];
+  writeNotNull(
+      'cameraDirection', _$CameraDirectionEnumMap[instance.cameraDirection]);
   return val;
 }
 
@@ -506,18 +537,15 @@ const _$CameraDirectionEnumMap = {
 
 ChannelMediaOptions _$ChannelMediaOptionsFromJson(Map<String, dynamic> json) {
   return ChannelMediaOptions(
-    json['autoSubscribeAudio'] as bool,
-    json['autoSubscribeVideo'] as bool,
-    publishLocalAudio: json['publishLocalAudio'] as bool,
-    publishLocalVideo: json['publishLocalVideo'] as bool,
+    autoSubscribeAudio: json['autoSubscribeAudio'] as bool?,
+    autoSubscribeVideo: json['autoSubscribeVideo'] as bool?,
+    publishLocalAudio: json['publishLocalAudio'] as bool?,
+    publishLocalVideo: json['publishLocalVideo'] as bool?,
   );
 }
 
 Map<String, dynamic> _$ChannelMediaOptionsToJson(ChannelMediaOptions instance) {
-  final val = <String, dynamic>{
-    'autoSubscribeAudio': instance.autoSubscribeAudio,
-    'autoSubscribeVideo': instance.autoSubscribeVideo,
-  };
+  final val = <String, dynamic>{};
 
   void writeNotNull(String key, dynamic value) {
     if (value != null) {
@@ -525,6 +553,8 @@ Map<String, dynamic> _$ChannelMediaOptionsToJson(ChannelMediaOptions instance) {
     }
   }
 
+  writeNotNull('autoSubscribeAudio', instance.autoSubscribeAudio);
+  writeNotNull('autoSubscribeVideo', instance.autoSubscribeVideo);
   writeNotNull('publishLocalAudio', instance.publishLocalAudio);
   writeNotNull('publishLocalVideo', instance.publishLocalVideo);
   return val;
@@ -532,18 +562,17 @@ Map<String, dynamic> _$ChannelMediaOptionsToJson(ChannelMediaOptions instance) {
 
 EncryptionConfig _$EncryptionConfigFromJson(Map<String, dynamic> json) {
   return EncryptionConfig(
-    _$enumDecodeNullable(_$EncryptionModeEnumMap, json['encryptionMode']),
-    json['encryptionKey'] as String,
-    encryptionKdfSalt:
-        (json['encryptionKdfSalt'] as List)?.map((e) => e as int)?.toList(),
+    encryptionMode:
+        _$enumDecodeNullable(_$EncryptionModeEnumMap, json['encryptionMode']),
+    encryptionKey: json['encryptionKey'] as String?,
+    encryptionKdfSalt: (json['encryptionKdfSalt'] as List<dynamic>?)
+        ?.map((e) => e as int)
+        .toList(),
   );
 }
 
 Map<String, dynamic> _$EncryptionConfigToJson(EncryptionConfig instance) {
-  final val = <String, dynamic>{
-    'encryptionMode': _$EncryptionModeEnumMap[instance.encryptionMode],
-    'encryptionKey': instance.encryptionKey,
-  };
+  final val = <String, dynamic>{};
 
   void writeNotNull(String key, dynamic value) {
     if (value != null) {
@@ -551,6 +580,9 @@ Map<String, dynamic> _$EncryptionConfigToJson(EncryptionConfig instance) {
     }
   }
 
+  writeNotNull(
+      'encryptionMode', _$EncryptionModeEnumMap[instance.encryptionMode]);
+  writeNotNull('encryptionKey', instance.encryptionKey);
   writeNotNull('encryptionKdfSalt', instance.encryptionKdfSalt);
   return val;
 }
@@ -568,30 +600,31 @@ const _$EncryptionModeEnumMap = {
 };
 
 RtcStats _$RtcStatsFromJson(Map<String, dynamic> json) {
-  return RtcStats()
-    ..duration = json['duration'] as int
-    ..txBytes = json['txBytes'] as int
-    ..rxBytes = json['rxBytes'] as int
-    ..txAudioBytes = json['txAudioBytes'] as int
-    ..txVideoBytes = json['txVideoBytes'] as int
-    ..rxAudioBytes = json['rxAudioBytes'] as int
-    ..rxVideoBytes = json['rxVideoBytes'] as int
-    ..txKBitRate = json['txKBitRate'] as int
-    ..rxKBitRate = json['rxKBitRate'] as int
-    ..txAudioKBitRate = json['txAudioKBitRate'] as int
-    ..rxAudioKBitRate = json['rxAudioKBitRate'] as int
-    ..txVideoKBitRate = json['txVideoKBitRate'] as int
-    ..rxVideoKBitRate = json['rxVideoKBitRate'] as int
-    ..userCount = json['userCount'] as int
-    ..lastmileDelay = json['lastmileDelay'] as int
-    ..txPacketLossRate = json['txPacketLossRate'] as int
-    ..rxPacketLossRate = json['rxPacketLossRate'] as int
-    ..cpuTotalUsage = (json['cpuTotalUsage'] as num)?.toDouble()
-    ..cpuAppUsage = (json['cpuAppUsage'] as num)?.toDouble()
-    ..gatewayRtt = json['gatewayRtt'] as int
-    ..memoryAppUsageRatio = (json['memoryAppUsageRatio'] as num)?.toDouble()
-    ..memoryTotalUsageRatio = (json['memoryTotalUsageRatio'] as num)?.toDouble()
-    ..memoryAppUsageInKbytes = json['memoryAppUsageInKbytes'] as int;
+  return RtcStats(
+    json['duration'] as int,
+    json['txBytes'] as int,
+    json['rxBytes'] as int,
+    json['txAudioBytes'] as int,
+    json['txVideoBytes'] as int,
+    json['rxAudioBytes'] as int,
+    json['rxVideoBytes'] as int,
+    json['txKBitRate'] as int,
+    json['rxKBitRate'] as int,
+    json['txAudioKBitRate'] as int,
+    json['rxAudioKBitRate'] as int,
+    json['txVideoKBitRate'] as int,
+    json['rxVideoKBitRate'] as int,
+    json['userCount'] as int,
+    json['lastmileDelay'] as int,
+    json['txPacketLossRate'] as int,
+    json['rxPacketLossRate'] as int,
+    (json['cpuTotalUsage'] as num).toDouble(),
+    (json['cpuAppUsage'] as num).toDouble(),
+    json['gatewayRtt'] as int,
+    (json['memoryAppUsageRatio'] as num).toDouble(),
+    (json['memoryTotalUsageRatio'] as num).toDouble(),
+    json['memoryAppUsageInKbytes'] as int,
+  );
 }
 
 Map<String, dynamic> _$RtcStatsToJson(RtcStats instance) => <String, dynamic>{
@@ -621,11 +654,12 @@ Map<String, dynamic> _$RtcStatsToJson(RtcStats instance) => <String, dynamic>{
     };
 
 AudioVolumeInfo _$AudioVolumeInfoFromJson(Map<String, dynamic> json) {
-  return AudioVolumeInfo()
-    ..uid = json['uid'] as int
-    ..volume = json['volume'] as int
-    ..vad = json['vad'] as int
-    ..channelId = json['channelId'] as String;
+  return AudioVolumeInfo(
+    json['uid'] as int,
+    json['volume'] as int,
+    json['vad'] as int,
+    json['channelId'] as String,
+  );
 }
 
 Map<String, dynamic> _$AudioVolumeInfoToJson(AudioVolumeInfo instance) =>
@@ -637,26 +671,37 @@ Map<String, dynamic> _$AudioVolumeInfoToJson(AudioVolumeInfo instance) =>
     };
 
 Rect _$RectFromJson(Map<String, dynamic> json) {
-  return Rect()
-    ..left = json['left'] as int
-    ..top = json['top'] as int
-    ..right = json['right'] as int
-    ..bottom = json['bottom'] as int;
+  return Rect(
+    left: json['left'] as int?,
+    top: json['top'] as int?,
+    right: json['right'] as int?,
+    bottom: json['bottom'] as int?,
+  );
 }
 
-Map<String, dynamic> _$RectToJson(Rect instance) => <String, dynamic>{
-      'left': instance.left,
-      'top': instance.top,
-      'right': instance.right,
-      'bottom': instance.bottom,
-    };
+Map<String, dynamic> _$RectToJson(Rect instance) {
+  final val = <String, dynamic>{};
+
+  void writeNotNull(String key, dynamic value) {
+    if (value != null) {
+      val[key] = value;
+    }
+  }
+
+  writeNotNull('left', instance.left);
+  writeNotNull('top', instance.top);
+  writeNotNull('right', instance.right);
+  writeNotNull('bottom', instance.bottom);
+  return val;
+}
 
 LastmileProbeOneWayResult _$LastmileProbeOneWayResultFromJson(
     Map<String, dynamic> json) {
-  return LastmileProbeOneWayResult()
-    ..packetLossRate = json['packetLossRate'] as int
-    ..jitter = json['jitter'] as int
-    ..availableBandwidth = json['availableBandwidth'] as int;
+  return LastmileProbeOneWayResult(
+    json['packetLossRate'] as int,
+    json['jitter'] as int,
+    json['availableBandwidth'] as int,
+  );
 }
 
 Map<String, dynamic> _$LastmileProbeOneWayResultToJson(
@@ -668,18 +713,14 @@ Map<String, dynamic> _$LastmileProbeOneWayResultToJson(
     };
 
 LastmileProbeResult _$LastmileProbeResultFromJson(Map<String, dynamic> json) {
-  return LastmileProbeResult()
-    ..state =
-        _$enumDecodeNullable(_$LastmileProbeResultStateEnumMap, json['state'])
-    ..rtt = json['rtt'] as int
-    ..uplinkReport = json['uplinkReport'] == null
-        ? null
-        : LastmileProbeOneWayResult.fromJson(
-            json['uplinkReport'] as Map<String, dynamic>)
-    ..downlinkReport = json['downlinkReport'] == null
-        ? null
-        : LastmileProbeOneWayResult.fromJson(
-            json['downlinkReport'] as Map<String, dynamic>);
+  return LastmileProbeResult(
+    _$enumDecode(_$LastmileProbeResultStateEnumMap, json['state']),
+    json['rtt'] as int,
+    LastmileProbeOneWayResult.fromJson(
+        json['uplinkReport'] as Map<String, dynamic>),
+    LastmileProbeOneWayResult.fromJson(
+        json['downlinkReport'] as Map<String, dynamic>),
+  );
 }
 
 Map<String, dynamic> _$LastmileProbeResultToJson(
@@ -687,8 +728,8 @@ Map<String, dynamic> _$LastmileProbeResultToJson(
     <String, dynamic>{
       'state': _$LastmileProbeResultStateEnumMap[instance.state],
       'rtt': instance.rtt,
-      'uplinkReport': instance.uplinkReport?.toJson(),
-      'downlinkReport': instance.downlinkReport?.toJson(),
+      'uplinkReport': instance.uplinkReport.toJson(),
+      'downlinkReport': instance.downlinkReport.toJson(),
     };
 
 const _$LastmileProbeResultStateEnumMap = {
@@ -698,11 +739,12 @@ const _$LastmileProbeResultStateEnumMap = {
 };
 
 LocalAudioStats _$LocalAudioStatsFromJson(Map<String, dynamic> json) {
-  return LocalAudioStats()
-    ..numChannels = json['numChannels'] as int
-    ..sentSampleRate = json['sentSampleRate'] as int
-    ..sentBitrate = json['sentBitrate'] as int
-    ..txPacketLossRate = json['txPacketLossRate'] as int;
+  return LocalAudioStats(
+    json['numChannels'] as int,
+    json['sentSampleRate'] as int,
+    json['sentBitrate'] as int,
+    json['txPacketLossRate'] as int,
+  );
 }
 
 Map<String, dynamic> _$LocalAudioStatsToJson(LocalAudioStats instance) =>
@@ -714,25 +756,25 @@ Map<String, dynamic> _$LocalAudioStatsToJson(LocalAudioStats instance) =>
     };
 
 LocalVideoStats _$LocalVideoStatsFromJson(Map<String, dynamic> json) {
-  return LocalVideoStats()
-    ..sentBitrate = json['sentBitrate'] as int
-    ..sentFrameRate = json['sentFrameRate'] as int
-    ..encoderOutputFrameRate = json['encoderOutputFrameRate'] as int
-    ..rendererOutputFrameRate = json['rendererOutputFrameRate'] as int
-    ..targetBitrate = json['targetBitrate'] as int
-    ..targetFrameRate = json['targetFrameRate'] as int
-    ..qualityAdaptIndication = _$enumDecodeNullable(
-        _$VideoQualityAdaptIndicationEnumMap, json['qualityAdaptIndication'])
-    ..encodedBitrate = json['encodedBitrate'] as int
-    ..encodedFrameWidth = json['encodedFrameWidth'] as int
-    ..encodedFrameHeight = json['encodedFrameHeight'] as int
-    ..encodedFrameCount = json['encodedFrameCount'] as int
-    ..codecType =
-        _$enumDecodeNullable(_$VideoCodecTypeEnumMap, json['codecType'])
-    ..txPacketLossRate = json['txPacketLossRate'] as int
-    ..captureFrameRate = json['captureFrameRate'] as int
-    ..captureBrightnessLevel = _$enumDecodeNullable(
-        _$CaptureBrightnessLevelTypeEnumMap, json['captureBrightnessLevel']);
+  return LocalVideoStats(
+    json['sentBitrate'] as int,
+    json['sentFrameRate'] as int,
+    json['encoderOutputFrameRate'] as int,
+    json['rendererOutputFrameRate'] as int,
+    json['targetBitrate'] as int,
+    json['targetFrameRate'] as int,
+    _$enumDecode(
+        _$VideoQualityAdaptIndicationEnumMap, json['qualityAdaptIndication']),
+    json['encodedBitrate'] as int,
+    json['encodedFrameWidth'] as int,
+    json['encodedFrameHeight'] as int,
+    json['encodedFrameCount'] as int,
+    _$enumDecode(_$VideoCodecTypeEnumMap, json['codecType']),
+    json['txPacketLossRate'] as int,
+    json['captureFrameRate'] as int,
+    _$enumDecode(
+        _$CaptureBrightnessLevelTypeEnumMap, json['captureBrightnessLevel']),
+  );
 }
 
 Map<String, dynamic> _$LocalVideoStatsToJson(LocalVideoStats instance) =>
@@ -777,24 +819,23 @@ const _$CaptureBrightnessLevelTypeEnumMap = {
 };
 
 RemoteAudioStats _$RemoteAudioStatsFromJson(Map<String, dynamic> json) {
-  return RemoteAudioStats()
-    ..uid = json['uid'] as int
-    ..quality = _$enumDecodeNullable(_$NetworkQualityEnumMap, json['quality'])
-    ..networkTransportDelay = json['networkTransportDelay'] as int
-    ..jitterBufferDelay = json['jitterBufferDelay'] as int
-    ..audioLossRate = json['audioLossRate'] as int
-    ..numChannels = json['numChannels'] as int
-    ..receivedSampleRate = json['receivedSampleRate'] as int
-    ..receivedBitrate = json['receivedBitrate'] as int
-    ..totalFrozenTime = json['totalFrozenTime'] as int
-    ..frozenRate = json['frozenRate'] as int
-    ..totalActiveTime = json['totalActiveTime'] as int
-    ..publishDuration = json['publishDuration'] as int
-    ..qoeQuality =
-        _$enumDecodeNullable(_$ExperienceQualityTypeEnumMap, json['qoeQuality'])
-    ..qualityChangedReason = _$enumDecodeNullable(
-        _$ExperiencePoorReasonEnumMap, json['qualityChangedReason'])
-    ..mosValue = json['mosValue'] as int;
+  return RemoteAudioStats(
+    json['uid'] as int,
+    _$enumDecode(_$NetworkQualityEnumMap, json['quality']),
+    json['networkTransportDelay'] as int,
+    json['jitterBufferDelay'] as int,
+    json['audioLossRate'] as int,
+    json['numChannels'] as int,
+    json['receivedSampleRate'] as int,
+    json['receivedBitrate'] as int,
+    json['totalFrozenTime'] as int,
+    json['frozenRate'] as int,
+    json['totalActiveTime'] as int,
+    json['publishDuration'] as int,
+    _$enumDecode(_$ExperienceQualityTypeEnumMap, json['qoeQuality']),
+    _$enumDecode(_$ExperiencePoorReasonEnumMap, json['qualityChangedReason']),
+    json['mosValue'] as int,
+  );
 }
 
 Map<String, dynamic> _$RemoteAudioStatsToJson(RemoteAudioStats instance) =>
@@ -843,21 +884,21 @@ const _$ExperiencePoorReasonEnumMap = {
 };
 
 RemoteVideoStats _$RemoteVideoStatsFromJson(Map<String, dynamic> json) {
-  return RemoteVideoStats()
-    ..uid = json['uid'] as int
-    ..delay = json['delay'] as int
-    ..width = json['width'] as int
-    ..height = json['height'] as int
-    ..receivedBitrate = json['receivedBitrate'] as int
-    ..decoderOutputFrameRate = json['decoderOutputFrameRate'] as int
-    ..rendererOutputFrameRate = json['rendererOutputFrameRate'] as int
-    ..packetLossRate = json['packetLossRate'] as int
-    ..rxStreamType =
-        _$enumDecodeNullable(_$VideoStreamTypeEnumMap, json['rxStreamType'])
-    ..totalFrozenTime = json['totalFrozenTime'] as int
-    ..frozenRate = json['frozenRate'] as int
-    ..totalActiveTime = json['totalActiveTime'] as int
-    ..publishDuration = json['publishDuration'] as int;
+  return RemoteVideoStats(
+    json['uid'] as int,
+    json['delay'] as int,
+    json['width'] as int,
+    json['height'] as int,
+    json['receivedBitrate'] as int,
+    json['decoderOutputFrameRate'] as int,
+    json['rendererOutputFrameRate'] as int,
+    json['packetLossRate'] as int,
+    _$enumDecode(_$VideoStreamTypeEnumMap, json['rxStreamType']),
+    json['totalFrozenTime'] as int,
+    json['frozenRate'] as int,
+    json['totalActiveTime'] as int,
+    json['publishDuration'] as int,
+  );
 }
 
 Map<String, dynamic> _$RemoteVideoStatsToJson(RemoteVideoStats instance) =>
@@ -883,12 +924,13 @@ const _$VideoStreamTypeEnumMap = {
 };
 
 FacePositionInfo _$FacePositionInfoFromJson(Map<String, dynamic> json) {
-  return FacePositionInfo()
-    ..x = json['x'] as int
-    ..y = json['y'] as int
-    ..width = json['width'] as int
-    ..height = json['height'] as int
-    ..distance = json['distance'] as int;
+  return FacePositionInfo(
+    json['x'] as int,
+    json['y'] as int,
+    json['width'] as int,
+    json['height'] as int,
+    json['distance'] as int,
+  );
 }
 
 Map<String, dynamic> _$FacePositionInfoToJson(FacePositionInfo instance) =>
@@ -902,16 +944,24 @@ Map<String, dynamic> _$FacePositionInfoToJson(FacePositionInfo instance) =>
 
 ClientRoleOptions _$ClientRoleOptionsFromJson(Map<String, dynamic> json) {
   return ClientRoleOptions(
-    _$enumDecodeNullable(
+    audienceLatencyLevel: _$enumDecodeNullable(
         _$AudienceLatencyLevelTypeEnumMap, json['audienceLatencyLevel']),
   );
 }
 
-Map<String, dynamic> _$ClientRoleOptionsToJson(ClientRoleOptions instance) =>
-    <String, dynamic>{
-      'audienceLatencyLevel':
-          _$AudienceLatencyLevelTypeEnumMap[instance.audienceLatencyLevel],
-    };
+Map<String, dynamic> _$ClientRoleOptionsToJson(ClientRoleOptions instance) {
+  final val = <String, dynamic>{};
+
+  void writeNotNull(String key, dynamic value) {
+    if (value != null) {
+      val[key] = value;
+    }
+  }
+
+  writeNotNull('audienceLatencyLevel',
+      _$AudienceLatencyLevelTypeEnumMap[instance.audienceLatencyLevel]);
+  return val;
+}
 
 const _$AudienceLatencyLevelTypeEnumMap = {
   AudienceLatencyLevelType.LowLatency: 1,
@@ -920,8 +970,8 @@ const _$AudienceLatencyLevelTypeEnumMap = {
 
 LogConfig _$LogConfigFromJson(Map<String, dynamic> json) {
   return LogConfig(
-    filePath: json['filePath'] as String,
-    fileSize: json['fileSize'] as int,
+    filePath: json['filePath'] as String?,
+    fileSize: json['fileSize'] as int?,
     level: _$enumDecodeNullable(_$LogLevelEnumMap, json['level']),
   );
 }
@@ -951,29 +1001,23 @@ const _$LogLevelEnumMap = {
 
 DataStreamConfig _$DataStreamConfigFromJson(Map<String, dynamic> json) {
   return DataStreamConfig(
-    syncWithAudio: json['syncWithAudio'] as bool,
-    ordered: json['ordered'] as bool,
+    json['syncWithAudio'] as bool,
+    json['ordered'] as bool,
   );
 }
 
-Map<String, dynamic> _$DataStreamConfigToJson(DataStreamConfig instance) {
-  final val = <String, dynamic>{};
-
-  void writeNotNull(String key, dynamic value) {
-    if (value != null) {
-      val[key] = value;
-    }
-  }
-
-  writeNotNull('syncWithAudio', instance.syncWithAudio);
-  writeNotNull('ordered', instance.ordered);
-  return val;
-}
+Map<String, dynamic> _$DataStreamConfigToJson(DataStreamConfig instance) =>
+    <String, dynamic>{
+      'syncWithAudio': instance.syncWithAudio,
+      'ordered': instance.ordered,
+    };
 
 RtcEngineConfig _$RtcEngineConfigFromJson(Map<String, dynamic> json) {
   return RtcEngineConfig(
     json['appId'] as String,
-    areaCode: _$enumDecodeNullable(_$AreaCodeEnumMap, json['areaCode']),
+    areaCode: (json['areaCode'] as List<dynamic>?)
+        ?.map((e) => _$enumDecode(_$AreaCodeEnumMap, e))
+        .toList(),
     logConfig: json['logConfig'] == null
         ? null
         : LogConfig.fromJson(json['logConfig'] as Map<String, dynamic>),
@@ -991,7 +1035,8 @@ Map<String, dynamic> _$RtcEngineConfigToJson(RtcEngineConfig instance) {
     }
   }
 
-  writeNotNull('areaCode', _$AreaCodeEnumMap[instance.areaCode]);
+  writeNotNull(
+      'areaCode', RtcEngineContext._$AreaCodeListToJson(instance.areaCode));
   writeNotNull('logConfig', instance.logConfig?.toJson());
   return val;
 }
@@ -1006,11 +1051,40 @@ const _$AreaCodeEnumMap = {
   AreaCode.GLOB: -1,
 };
 
+RtcEngineContext _$RtcEngineContextFromJson(Map<String, dynamic> json) {
+  return RtcEngineContext(
+    json['appId'] as String,
+    areaCode: (json['areaCode'] as List<dynamic>?)
+        ?.map((e) => _$enumDecode(_$AreaCodeEnumMap, e))
+        .toList(),
+    logConfig: json['logConfig'] == null
+        ? null
+        : LogConfig.fromJson(json['logConfig'] as Map<String, dynamic>),
+  );
+}
+
+Map<String, dynamic> _$RtcEngineContextToJson(RtcEngineContext instance) {
+  final val = <String, dynamic>{
+    'appId': instance.appId,
+  };
+
+  void writeNotNull(String key, dynamic value) {
+    if (value != null) {
+      val[key] = value;
+    }
+  }
+
+  writeNotNull(
+      'areaCode', RtcEngineContext._$AreaCodeListToJson(instance.areaCode));
+  writeNotNull('logConfig', instance.logConfig?.toJson());
+  return val;
+}
+
 RhythmPlayerConfig _$RhythmPlayerConfigFromJson(Map<String, dynamic> json) {
   return RhythmPlayerConfig(
-    beatsPerMeasure: json['beatsPerMeasure'] as int,
-    beatsPerMinute: json['beatsPerMinute'] as int,
-    publish: json['publish'] as bool,
+    beatsPerMeasure: json['beatsPerMeasure'] as int?,
+    beatsPerMinute: json['beatsPerMinute'] as int?,
+    publish: json['publish'] as bool?,
   );
 }